# src/main/resources/openapi.yml
openapi: 3.0.1
info:
  title: Teensy
  description: Tiny Url Generator
  version: 1.0.0
  contact:
    name: Tejas Ozarkar
    email: tejozarkar@gmail.com
servers:
  - url: http://localhost:8080
    description: Local Environment
  - url: http://teensyapi-env.eba-k2xfeapv.ap-south-1.elasticbeanstalk.com
    description: Production Environment
paths:
...
  /card:
    post:
      summary: Crate New Card
      operationId: createCard
      requestBody:
        description: Create Card
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardRequest'
        required: true
      responses:
        200:
          description: New Card as Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
    get:
      summary: Get all cards of user
      operationId: getUserCards
      responses:
        200:
          description: Get user cards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CardResponse'
  /card/{cardId}:
    patch:
      summary: Edit the Card
      operationId: editCard
      requestBody:
        description: Edit Card
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCardDto'
        required: true
      responses:
        200:
          description: Edit the card
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
      parameters:
        - name: cardId
          in: path
          description: Id of card
          required: true
          schema:
            type: integer
    delete:
      summary: Delete selected card
      operationId: deleteCard()
      responses:
        200:
          description: Delete Card
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
      parameters:
        - name: cardId
          in: path
          description: Id of card
          required: true
          schema:
            type: integer
    get:
      summary: Get Card Info
      operationId: getCardInfo
      responses:
        200:
          description: Get Card Info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
      parameters:
        - name: cardId
          in: path
          description: Id of card
          required: true
          schema:
            type: integer
  /card/bygroup/{groupId}:
    get:
      summary: Delete selected card
      operationId: getGroupCards
      responses:
        200:
          description: Get Group Cards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CardResponse'
      parameters:
        - name: groupId
          in: path
          description: Id of group
          required: true
          schema:
            type: integer
...
components:
  schemas:
...
    UrlResponseDto:
      type: object
      properties:
        id:
          type: integer
        shorturl:
          type: string
        longurl:
          type: string
        expirydate:
          type: string
...
    CardDto:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        icon:
          type: string
    UpdateCardDto:
      allOf:
        - $ref: '#/components/schemas/CardDto'
      type: object
      properties:
        longurl:
          type: string
    CardRequest:
      allOf:
        - $ref: '#/components/schemas/CardDto'
      type: object
      properties:
        longurl:
          type: string
        groupid:
          type: integer
    CardResponse:
      allOf:
        - $ref: '#/components/schemas/CardRequest'
        - $ref: '#/components/schemas/UrlResponseDto'
      type: object
      properties:
        id:
          type: integer
        urlid:
          type: integer
        groupid:
          type: integer
        grouptitle:
          type: string
        groupdescription:
          type: string
...
    GenericResponse:
      type: object
      properties:
        message:
          type: string